fix: wire artifact remover into all CLI entrypoints; DRY _build_artifact_remover
- Move _build_artifact_remover(seg_cfg: dict) to tessellate.py (co-located
  with SegConfig which defines remove_artifacts/remove_penmarks/artifact_exclude_classes)
- Wire artifact_remover_fn into tessellate.py::main() and filter_tessellate.py::_main()
  so all three CLI entry points (tessellate, filter-tessellate,
  tessellate-extract-features) consistently apply artifact removal
- Import _build_artifact_remover in filter_tessellate, tessellate_extract_features,
  and tessellate_extract_features_common instead of duplicating the logic
- Replace inline artifact remover construction in tessellate_extract_features_common
  with a single call to _build_artifact_remover
- Remove unused call import from test_tessellate.py; reformat with black;
  update GrandQCArtifactRemover patch target to mussel.cli.tessellate
