+++ mussel/cli/tessellate_extract_features_common.py
@@ -12,14 +12,11 @@ from shapely.geometry import Polygon
 
 logger = logging.getLogger(__name__)
 
+from mussel.cli.tessellate import _build_artifact_remover
 from mussel.utils import (filter_features, is_remote_path, load_classifier,
                           load_features_from_h5, safe_path_join, save_features,
                           save_hdf5, save_torch_tensor)
-from mussel.utils.artifact_removal import (
-    EXCLUDE_ALL_ARTIFACTS,
-    EXCLUDE_PENMARKS_ONLY,
-    GrandQCArtifactRemover,
-)
+from mussel.utils.artifact_removal import GrandQCArtifactRemover
 from mussel.utils.segment import (draw_slide_mask, save_patches_png,
                                   segment_tissue)
 
@@ -86,22 +83,7 @@ def _tessellate_and_filter(
     seg_cfg = OmegaConf.to_container(cfg.seg_config)
 
     if artifact_remover_fn is None:
-        if seg_cfg.get("remove_artifacts") or seg_cfg.get("remove_penmarks"):
-            explicit = seg_cfg.get("artifact_exclude_classes")
-            if explicit:
-                exclude_classes = frozenset(int(c) for c in explicit)
-            elif bool(seg_cfg.get("remove_artifacts")):
-                exclude_classes = EXCLUDE_ALL_ARTIFACTS
-            else:
-                exclude_classes = EXCLUDE_PENMARKS_ONLY
-            artifact_remover_fn = GrandQCArtifactRemover(exclude_classes=exclude_classes)
-            logger.info(
-                "Artifact removal enabled: remove_artifacts=%s remove_penmarks=%s "
-                "exclude_classes=%s",
-                seg_cfg.get("remove_artifacts"),
-                seg_cfg.get("remove_penmarks"),
-                sorted(exclude_classes),
-            )
+        artifact_remover_fn = _build_artifact_remover(seg_cfg)
 
     # Strip config-only keys that are not segment_tissue() parameters.
     seg_cfg.pop("artifact_exclude_classes", None)
